return errors via `Error` enum
